Fixes #8702. Move error registry disabled test classes to their own files
